+++ src/webaudio.js
@@ -447,6 +447,12 @@ export default class WebAudio extends util.Observer {
             for (i = first; i <= last; i++) {
                 const start = ~~(i * sampleSize);
                 const end = ~~(start + sampleSize);
+		/**
+		 * Initialize the max and min to the first sample of this
+		 * subrange, so that even if the samples are entirely
+		 * on one side of zero, we still return the true max and
+		 * min values in the subrange.
+		 */
                 let min = chan[start];
                 let max = min;
                 let j;