+++ CHANGES.md
@@ -26,7 +26,7 @@ wavesurfer.js changelog
   - Fixed switch loop region (#1929)
   - Added ability to specify time format for Regions tooltip using timeformatCallback (#1948)
 - Add `splitChannelsOptions` param and `setFilteredChannels` method to configure how channels are drawn (#1947)
-- Added checks in `minimap` plugin for `drawer` presence
+- Added checks in `minimap` plugin for `drawer` presence (#1953)
 
 3.3.3 (16.04.2020)
 ------------------