+++ CHANGES.md
@@ -28,6 +28,7 @@ wavesurfer.js changelog
 - Add `splitChannelsOptions` param and `setFilteredChannels` method to configure how channels are drawn (#1947)
 - Added checks in `minimap` plugin for `drawer` presence (#1953)
 - Add `setDisabledEventEmissions` method to optionally disable calls to event handlers for specific events (#1960)
+- Drawer: removed private methods to allow inheriting them (#1962)
 
 3.3.3 (16.04.2020)
 ------------------